% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/poi.R
\docType{data}
\name{poi}
\alias{poi}
\title{poi: Points of interest for waste collection in Blantyre, Malawi}
\format{
A tibble with 5 rows and 3 variables:
\describe{
\item{poi}{Point of interest.}
\item{lat}{Y (latitude) coordinate in WGS 84.}
\item{long}{X (longitude) coordinate in WGS 84.}
}
}
\usage{
poi
}
\description{
Points of interest for the collection of waste from the public skips in
Blantyre, Malawi: the depot, the Mzedi dump site, the compost site and
two gas stations.
}
\keyword{datasets}
