% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wasteskipsblantyre.R
\docType{data}
\name{wasteskipsblantyre}
\alias{wasteskipsblantyre}
\title{Location of waste skips in Blantyre, Malawi}
\format{
A tibble with 53 rows and 4 variables:
\describe{
\item{name}{Name of the waste skip location.}
\item{long}{Longitude of the location as decimal degrees.}
\item{lat}{Latitude of the location as decimal degrees.}
\item{capacity_l}{Capacity of the container in liters.}
}
}
\usage{
wasteskipsblantyre
}
\description{
This data set includes spatial data about the location of waste skips
in the city of Blantyre, Malawi. The data was collected in 2021.
}
\keyword{datasets}
