% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wasteskipsblantyre-package.R
\docType{package}
\name{wasteskipsblantyre-package}
\alias{wasteskipsblantyre-package}
\alias{_PACKAGE}
\title{wasteskipsblantyre: Locations of Public Waste Skips in Blantyre, Malawi}
\description{
An R data package containing the locations of public waste skips in Blantyre, Malawi.
}
\author{
\strong{Maintainer}: Sebastian Camilo Loos \email{seba.loos@hotmail.com} (\href{https://orcid.org/0000-0002-8830-1734}{ORCID})

Authors:
\itemize{
  \item Mabvuto Yesaya \email{yesayamabvuto@gmail.com} (\href{https://orcid.org/0000-0003-3234-9699}{ORCID})
  \item Limbani Msuku \email{limbani_msuku@yahoo.com} (\href{https://orcid.org/0000-0002-9291-8462}{ORCID})
  \item Elizabeth Tilley \email{tilleye@ethz.ch} (\href{https://orcid.org/0000-0002-2095-9724}{ORCID})
}

Other contributors:
\itemize{
  \item Lars Schöbitz \email{lschoebitz@ethz.ch} (\href{https://orcid.org/0000-0003-2196-5015}{ORCID}) [contributor]
  \item Mian Zhong \email{mzhong@ethz.ch} (\href{https://orcid.org/0009-0009-4546-7214}{ORCID}) [contributor]
}

}
\keyword{internal}
